requirements-txt.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-login==0.6.2
flask-wtf==1.1.1
werkzeug==2.3.7
pymysql==1.1.0
nltk==3.8.1
numpy==1.25.2
pandas==2.1.0
PyPDF2==3.0.1
transformers==4.36.2
sklearn==0.0.post10
torch==2.1.0
python-dotenv==1.0.0
